Add api.d.ts
